backend/requirements.txt
asgiref==3.7.2
axios==0.4.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
Django==4.2.2
django-cors-headers==4.1.0
django-environ==0.10.0
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
lxml==4.9.3
markdown-it-py==3.0.0
mdurl==0.1.2
Pillow==9.5.0
psycopg2==2.9.6
Pygments==2.16.1
PyJWT==2.8.0
pytz==2023.3
requests==2.31.0
rich==13.5.2
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
